backend/requirements.txt
Django==4.1.2
PyMySQL==1.0.2
django-admin-interface==0.24.2
django-import-export==2.9.0
djangorestframework==3.13.1
django-reversion==5.0.3
django-redis==5.2.0
djangorestframework-simplejwt==5.2.2
django-filter==21.1
django-celery-beat==2.4.0
django-celery-results==2.4.0
celery==5.2.7
ipaddr==2.2.0
flower==1.2.0
reversion==0.2
openwisp-users==1.0.2
openwisp-utils==1.0.4
django-allauth==0.51.0
django-cidrfield==0.2.0
netaddr==0.7.19
pandas==1.5.0
pymongo==3.10.1
pycryptodome==3.15.0
supervisor==4.2.4
daphne==2.5.0